Integrate webpack build and remove Browserify
Allows easier builds without the need for ReactProvider and D3Provider
files. Also included new minified build w/ sourcemap for browser users
of the library.

Also renamed all .js files to .jsx where appropriate and updated to the
latest package dependencies.
