+++ .gitignore
@@ -1,1 +1,2 @@
+lib/
 node_modules/
