+++ package.json
@@ -1,42 +1,39 @@
 {
-	"name": "react-d3-components",
-	"version": "0.3.4",
-	"description": "D3 components for React",
-	"main": "src/index.js",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/codesuki/react-d3-components.git"
-	},
-	"keywords": [
-		"d3",
-		"react",
-		"graph",
-		"visualization",
-		"chart",
-		"react-component"
-	],
-	"author": "Neri Marschik <marschik_neri@cyberagent.co.jp> (http://www.cyberagent.co.jp)",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/codesuki/react-d3-components/issues"
-	},
-	"homepage": "https://github.com/codesuki/react-d3-components",
-	"dependencies": {
-		"d3": "^3.5.3",
-		"react": "^0.12.2"
-	},
-	"devDependencies": {
-		"6to5ify": "^4.1.0",
-		"browserify": "^8.1.3",
-		"reactify": "^1.0.0"
-	},
-	"scripts": {
-		"build:js": "browserify src/index.js --no-bundle-external --standalone ReactD3 > dist/react-d3-components.js"
-	},
-	"browserify": {
-		"transform": [
-			"reactify",
-			"6to5ify"
-		]
-	}
+  "name": "react-d3-components",
+  "version": "0.3.4",
+  "description": "D3 components for React",
+  "main": "lib/index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/codesuki/react-d3-components.git"
+  },
+  "keywords": [
+    "d3",
+    "react",
+    "graph",
+    "visualization",
+    "chart",
+    "react-component"
+  ],
+  "author": "Neri Marschik <marschik_neri@cyberagent.co.jp> (http://www.cyberagent.co.jp)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/codesuki/react-d3-components/issues"
+  },
+  "homepage": "https://github.com/codesuki/react-d3-components",
+  "dependencies": {
+    "babel": "^4.7.12",
+    "d3": "^3.5.5",
+    "react": "^0.12.2"
+  },
+  "devDependencies": {
+    "babel-core": "^4.7.12",
+    "babel-loader": "^4.1.0",
+    "webpack": "^1.7.3"
+  },
+  "scripts": {
+    "compile": "babel -d lib/ src/",
+    "prepublish": "npm run compile",
+    "build:js": "webpack && webpack --config webpack.config.min.js"
+  }
 }
