+++ src/BarChart.jsx
@@ -1,86 +1,83 @@
-let React = require('./ReactProvider');
-let d3 = require('./D3Provider');
+let React = require('react');
+let d3 = require('d3');
 
 let Chart = require('./Chart');
 let Axis = require('./Axis');
+let Bar = require('./Bar');
 let Tooltip = require('./Tooltip');
 
 let DefaultPropsMixin = require('./DefaultPropsMixin');
 let HeightWidthMixin = require('./HeightWidthMixin');
 let ArrayifyMixin = require('./ArrayifyMixin');
-let AccessorMixin = require('./AccessorMixin');
+let StackAccessorMixin = require('./StackAccessorMixin');
+let StackDataMixin = require('./StackDataMixin');
 let DefaultScalesMixin = require('./DefaultScalesMixin');
 let TooltipMixin = require('./TooltipMixin');
 
 let DataSet = React.createClass({
 	propTypes: {
 		data: React.PropTypes.array.isRequired,
-		symbol: React.PropTypes.func.isRequired,
 		xScale: React.PropTypes.func.isRequired,
 		yScale: React.PropTypes.func.isRequired,
 		colorScale: React.PropTypes.func.isRequired,
-		onMouseEnter: React.PropTypes.func,
-		onMouseLeave: React.PropTypes.func
+		values: React.PropTypes.func.isRequired,
+		label: React.PropTypes.func.isRequired,
+		x: React.PropTypes.func.isRequired,
+		y: React.PropTypes.func.isRequired,
+		y0: React.PropTypes.func.isRequired
 	},
 
 	render() {
 		let {data,
-			 symbol,
 			 xScale,
 			 yScale,
 			 colorScale,
 			 values,
+			 label,
 			 x,
 			 y,
+			 y0,
 			 onMouseEnter,
 			 onMouseLeave} = this.props;
 
-		let circles = data.map(stack => {
+		let bars = data.map(stack => {
 			return values(stack).map(e => {
-				let translate = `translate(${xScale(x(e))}, ${yScale(y(e))})`;
 				return (
-						<path
-					className="dot"
-					d={symbol()}
-					transform={translate}
-					fill={colorScale(stack.label)}
-					onMouseOver={ evt => { onMouseEnter(evt, e); } }
-					onMouseLeave={  evt => { onMouseLeave(evt); } }
+						<Bar
+					width={xScale.rangeBand()}
+					height={yScale(yScale.domain()[0]) - yScale(y(e))}
+					x={xScale(x(e))}
+					y={yScale(y0(e) + y(e))}
+					fill={colorScale(label(stack))}
+					data={e}
+					onMouseEnter={onMouseEnter}
+					onMouseLeave={onMouseLeave}
 						/>
 				);
 			});
 		});
 
 		return (
-				<g>
-				{circles}
-			</g>
+				<g>{bars}</g>
 		);
 	}
 });
 
-let ScatterPlot = React.createClass({
+let BarChart = React.createClass({
 	mixins: [DefaultPropsMixin,
 			 HeightWidthMixin,
 			 ArrayifyMixin,
-			 AccessorMixin,
+			 StackAccessorMixin,
+			 StackDataMixin,
 			 DefaultScalesMixin,
 			 TooltipMixin],
 
-	propTypes: {
-		rScale: React.PropTypes.func,
-		shape: React.PropTypes.string
-	},
-
 	getDefaultProps() {
-		return {
-			rScale: null,
-			shape: 'circle'
-		};
+		return {};
 	},
 
 	_tooltipHtml(d, position) {
-		return this.props.tooltipHtml(this.props.x(d), this.props.y(d));
+		return this.props.tooltipHtml(this.props.x(d), this.props.y0(d), this.props.y(d));
 	},
 
 	render() {
@@ -88,11 +85,11 @@ let ScatterPlot = React.createClass({
 			 width,
 			 margin,
 			 colorScale,
-			 rScale,
-			 shape,
 			 values,
-			 x,
+			 label,
 			 y,
+			 y0,
+			 x,
 			 xAxis,
 			 yAxis} = this.props;
 
@@ -100,56 +97,45 @@ let ScatterPlot = React.createClass({
 			 innerWidth,
 			 innerHeight,
 			 xScale,
-			 yScale,
-			 xIntercept,
-			 yIntercept] = [this._data,
-							this._innerWidth,
-							this._innerHeight,
-							this._xScale,
-							this._yScale,
-							this._xIntercept,
-							this._yIntercept];
-
-		let symbol = d3.svg.symbol().type(shape);
-
-		if (rScale) {
-			symbol = symbol.size(rScale);
-		}
+			 yScale] = [this._data,
+						this._innerWidth,
+						this._innerHeight,
+						this._xScale,
+						this._yScale];
 
 		return (
-			<div>
+				<div>
 				<Chart height={height} width={width} margin={margin}>
-				<Axis
-			className={"x axis"}
-			orientation="bottom"
-			scale={xScale}
-			height={innerHeight}
-			width={innerWidth}
-			zero={yIntercept}
-			{...xAxis}
-				/>
-
-				<Axis
-			className={"y axis"}
-			orientation="left"
-			scale={yScale}
-			height={innerHeight}
-			width={innerWidth}
-			zero={xIntercept}
-			{...yAxis}
-				/>
-
 				<DataSet
 			data={data}
 			xScale={xScale}
 			yScale={yScale}
 			colorScale={colorScale}
-			symbol={symbol}
 			values={values}
-			x={x}
+			label={label}
 			y={y}
+			y0={y0}
+			x={x}
 			onMouseEnter={this.onMouseEnter}
 			onMouseLeave={this.onMouseLeave}
+				/>
+
+				<Axis
+			className={"x axis"}
+			orientation={"bottom"}
+			scale={xScale}
+			height={innerHeight}
+			width={innerWidth}
+			{...xAxis}
+				/>
+
+				<Axis
+			className={"y axis"}
+			orientation={"left"}
+			scale={yScale}
+			height={innerHeight}
+			width={innerWidth}
+			{...yAxis}
 				/>
 				</Chart>
 
@@ -163,4 +149,4 @@ let ScatterPlot = React.createClass({
 	}
 });
 
-module.exports = ScatterPlot;
+module.exports = BarChart;
