--- src/D3Provider.js
@@ -1,3 +1,0 @@
-var d3 = window.d3 || require('./D3Provider');
-
-module.exports = d3;
