+++ src/PieChart.jsx
@@ -1,5 +1,5 @@
-let React = require('./ReactProvider');
-let d3 = require('./D3Provider');
+let React = require('react');
+let d3 = require('d3');
 
 let Chart = require('./Chart');
 let Tooltip = require('./Tooltip');
