--- src/ReactProvider.js
@@ -1,3 +1,0 @@
-var React = window.React || require('react');
-
-module.exports = React;
